feat(app): support bind pid (#6)
When the bound pid exits, the ovm also needs to exit.

Because there is a possibility that the app sets detached when launching
ovm, at this time ppid will be 1 (ovm becomes a zombie process). In
order to avoid this situation, we need users to explicitly pass in the
current app's pid. Avoid the problem of residual processes.

Signed-off-by: Black-Hole1 <bh@bugs.cc>