feat(vm): sync time when vm state is resume
After the virtual machine resumes from the pause, we need to synchronize the time, as the time in the VM is inaccurate now.
Here, we are listening to `VirtualMachineStateResuming`, as there's no `VirtualMachineStateResume` state.
And the VM progresses from `VirtualMachineStateResuming` to `VirtualMachineStateRunning` quite rapidly, even if it's slow, the request to synchronize time won't fail but will be piled up at the sender's end, until the other end recovers.

Signed-off-by: Black-Hole1 <bh@bugs.cc>